AV mthfkn X!

+++ mandelbrot.h
@@ -1,2 +1,2 @@
 int getIterations(float x, float y, int limit);
-void iterations(float* x, float y, int limit, int* result);
+void iterations(float* x, float y, float limit, float* result);
